Renamed directories based on Oli's suggestion. Should be the final reorganization commit.
