+++ rir/src/analysis/dataflow.h
@@ -3,8 +3,8 @@
 
 #include "R/Funtab.h"
 #include "R/r.h"
-#include "analysis/analysis.h"
-#include "analysis/dispatchers.h"
+#include "analysis_framework/analysis.h"
+#include "analysis_framework/dispatchers.h"
 #include "interpreter/interp_context.h"
 #include "ir/CodeEditor.h"
