+++ rir/src/analysis_framework/analysis.h
