+++ rir/src/analysis_framework/framework.h
